chore: rename classes
