+++ core/android/src/main/kotlin/com/walletconnect/android/archive/HistoryInterface.kt
@@ -1,7 +1,7 @@
-package com.walletconnect.android.history
+package com.walletconnect.android.archive
 
 import com.walletconnect.android.Core
-import com.walletconnect.android.history.network.model.messages.MessagesParams
+import com.walletconnect.android.archive.network.model.messages.MessagesParams
 import com.walletconnect.android.internal.common.model.HistoryMessage
 import com.walletconnect.android.internal.common.model.Tags
