+++ core/android/src/main/kotlin/com/walletconnect/android/archive/HistoryMessageNotifier.kt
@@ -1,4 +1,4 @@
-package com.walletconnect.android.history
+package com.walletconnect.android.archive
 
 import com.walletconnect.foundation.network.model.Relay
 import kotlinx.coroutines.flow.MutableSharedFlow
