+++ core/android/src/main/kotlin/com/walletconnect/android/archive/HistoryProtocol.kt
@@ -1,9 +1,9 @@
-package com.walletconnect.android.history
+package com.walletconnect.android.archive
 
 import com.walletconnect.android.Core
-import com.walletconnect.android.history.domain.GetMessagesUseCase
-import com.walletconnect.android.history.domain.RegisterTagsUseCase
-import com.walletconnect.android.history.network.model.messages.MessagesParams
+import com.walletconnect.android.archive.domain.GetMessagesUseCase
+import com.walletconnect.android.archive.domain.RegisterTagsUseCase
+import com.walletconnect.android.archive.network.model.messages.MessagesParams
 import com.walletconnect.android.internal.common.di.AndroidCommonDITags
 import com.walletconnect.android.internal.common.model.HistoryMessage
 import com.walletconnect.android.internal.common.model.Tags
