+++ core/android/src/main/kotlin/com/walletconnect/android/archive/domain/RegisterTagsUseCase.kt
@@ -1,7 +1,7 @@
-package com.walletconnect.android.history.domain
+package com.walletconnect.android.archive.domain
 
-import com.walletconnect.android.history.network.HistoryServerService
-import com.walletconnect.android.history.network.model.register.RegisterBody
+import com.walletconnect.android.archive.network.HistoryServerService
+import com.walletconnect.android.archive.network.model.register.RegisterBody
 import com.walletconnect.android.internal.common.jwt.clientid.GenerateJwtStoreClientIdUseCase
 import com.walletconnect.android.internal.common.model.Tags
 import com.walletconnect.foundation.util.Logger
