+++ core/android/src/main/kotlin/com/walletconnect/android/archive/network/HistoryServerService.kt
@@ -1,8 +1,8 @@
-package com.walletconnect.android.history.network
+package com.walletconnect.android.archive.network
 
-import com.walletconnect.android.history.network.model.messages.MessagesResponse
-import com.walletconnect.android.history.network.model.register.RegisterBody
-import com.walletconnect.android.history.network.model.register.RegisterResponse
+import com.walletconnect.android.archive.network.model.messages.MessagesResponse
+import com.walletconnect.android.archive.network.model.register.RegisterBody
+import com.walletconnect.android.archive.network.model.register.RegisterResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
