+++ core/android/src/main/kotlin/com/walletconnect/android/internal/common/json_rpc/domain/JsonRpcInteractor.kt
@@ -1,6 +1,6 @@
 package com.walletconnect.android.internal.common.json_rpc.domain
 
-import com.walletconnect.android.history.HistoryMessageNotifier
+import com.walletconnect.android.archive.HistoryMessageNotifier
 import com.walletconnect.android.internal.common.JsonRpcResponse
 import com.walletconnect.android.internal.common.crypto.codec.Codec
 import com.walletconnect.android.internal.common.exception.NoRelayConnectionException
